Оказалось, что трансивер СС1120 проблематично настроить на длину пакетов более 128 байт. Поэтому пришлось отказаться от применения кодера 1/2 к данным. В противном случае пришлось бы урезать максимальный размер полезной нагрузки пакетов до 64 байт, а это менее ранее определенного стандартного 81-байтного размера, используемого как в речевом, так и информационнном обмене. В общем, пришлось дореализовать еще один кодер с эффективностью 3/4. Алгоритм также заимствован из стандарта APCO25, хоть и не очень предназначен для 2-позиционной модуляции. Кроме непосредственной реализации функций помехоустойчивого кодирования сделано следующее: - В функции формирования и обработки радиоосообщений (класс RadioMessage) добавлена возможность формирования и обработки сообщений произвольного размера (размер стал входных параметром) - Обработчик состояний радиоинтерфейса ProcessRadioState() разделен на 4 функции в виду того, что сильно быстро рос

+++ Src/CrossPlatform/cmx7262.c
@@ -498,7 +498,7 @@ void CMX7262_Test_AudioOut (CMX7262_TypeDef *pCmx7262)
 	
 	//Код частоты NCO
 	//Расчет кода частоты - п.8.1.9 Frequency Control документа D/7262_FI-1.x/4 August 2013 (datasheet)
-	uData = floor((nFreq * UINT16_MAX)/CMX7262_FREQ_SAMPLING);	
+	uData = floor(((double)nFreq * UINT16_MAX)/CMX7262_FREQ_SAMPLING);
 	CBUS_Write16(FREQ_CONTROL,&uData,1,pCmx7262->uInterface);
 	
 	// The test mode is started, there will be a delay before we are requested to service it..
